requirements.txt
Django==4.2.5
pytest-django==4.5.2
pytest-lazy-fixture==0.6.3
